% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialise.R
\name{autocomplete}
\alias{autocomplete}
\title{Autocomplete in Forms}
\usage{
autocomplete(id, inputIcon, label, datalist)
}
\arguments{
\item{id}{A character.}

\item{inputIcon}{A tag that shows icon on the left of the input field.}

\item{label}{A charater, showing the label in input area.}

\item{datalist}{A list of name and value pairs. Names will be autocomplete for input field}
}
\value{

}
\description{
Autocomplete in Forms
}
\examples{
inputIcon <- tags$i(class = "material-icons prefix",
           "textsms")
id ="autocomplete-input"
label = "Autocomplete"
list(
  apple=NULL, microsoft=NULL, google='https://placehold.it/250x250'
) -> datalist
tag_autocomplete <- autocomplete(
  id=id, inputIcon=inputIcon, label=label, datalist=datalist
)
}
