% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialise.R
\name{element_nav}
\alias{element_nav}
\title{Generate li tags for navbar}
\usage{
element_nav(href, label)
}
\arguments{
\item{href}{A href link}

\item{label}{A label}
}
\value{

}
\description{
Generate li tags for navbar
}
\examples{
none
}
