% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialise.R
\name{element_tab}
\alias{element_tab}
\title{Element function for tabs}
\usage{
element_tab(id, name, content, active = F, disabled = F)
}
\arguments{
\item{id}{An id for the tab}

\item{name}{A name for the tab}

\item{content}{A tags class object or a character}

\item{active}{logical, default=F}

\item{disabled}{logical, default=F}
}
\value{

}
\description{
Element function for tabs
}
\examples{
none
}
