% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialise.R
\name{pageLayout}
\alias{pageLayout}
\title{Simple layout}
\usage{
pageLayout(
  sidebarPanel = "",
  mainPanel = "",
  width_sidebar = "s3",
  width_main = "s9",
  class_sidebar = "sidebar",
  class_main = "main"
)
}
\arguments{
\item{sidebarPanel}{A tag object. (width fixed at s3)}

\item{mainPanel}{A tag object. (width fixed at s9)}

\item{width_sidebar}{a character descripting the column width following materialise css}

\item{width_main}{a character descripting the column width following materialise css}

\item{class_sidebar}{extra class name for sidebar}

\item{class_main}{extra class name for main}
}
\value{

}
\description{
Simple layout
}
\examples{
none
}
