% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialise.R
\name{tabs}
\alias{tabs}
\title{Generate tabs}
\usage{
tabs(...)
}
\arguments{
\item{...}{inputs generated by element_tab, separated by comma}
}
\value{

}
\description{
Generate tabs
}
\examples{
tabs(element_tab(...), element_tab(...))
}
